Move tools into a tools/ folder
